refactor(gateway): dependency-injected config via createGateway factory (MYM-26)
Replace the module-load-frozen `gwEnv` IIFE with injected config so no source
module reads global `Bun.env` except the entrypoint. This removes the test
coupling to mutable process state that caused sign/verify token mismatches.

- config.ts (was env.ts): `loadConfigFromEnv(env): GatewayConfig` parses and
  validates once, returning a typed config.
- db.ts: `getDb()`/singleton → `createDb(databaseUrl): Db`; no env import.
- gateway.ts (new): `createGateway(config, db): Hono` — pure, config in / app
  out. Routes + shared `bearerClaims(c, secret)` moved here.
- index.ts: thin entrypoint, the only module that reads `Bun.env`.
- Tests construct the app explicitly and sign with `config.llmTokenSecret`,
  making the sign/verify mismatch unrepresentable. `setDbForTests`/`testDb`
  removed for constructor injection; `test-setup.ts` + `bunfig.toml` deleted.
