+++ apps/gateway/src/db.ts
@@ -1,5 +1,4 @@
 import { SQL } from "bun";
-import { gwEnv } from "./env";
 
 /**
  * Minimal data-access seam. Parameterized SQL only — every caller passes a
@@ -13,16 +12,12 @@ export interface Db {
 	): Promise<T[]>;
 }
 
-let singleton: Db | undefined;
-
-/** The real Db, backed by Bun's built-in Postgres client (lazy, pooled). */
-export function getDb(): Db {
-	if (singleton) return singleton;
+/** The real Db, backed by Bun's built-in Postgres client (pooled). */
+export function createDb(databaseUrl: string): Db {
 	// TLS is carried in the URL (sslmode); read-only gateway → a small pool.
-	const sql = new SQL({ url: gwEnv.DATABASE_URL, max: 8 });
-	singleton = {
+	const sql = new SQL({ url: databaseUrl, max: 8 });
+	return {
 		query: <T>(text: string, params: unknown[] = []) =>
 			sql.unsafe(text, params) as Promise<T[]>,
 	};
-	return singleton;
 }
